Se removieron los archivos java main innecesarios, y se arreglaron algunos errores de localizaciones
