+++ nbproject/private/private.xml
@@ -3,12 +3,10 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/gui/JFrameMcQuinnator.java</file>
-            <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/gui/JFrameStarterPack.java</file>
             <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/McQuinnator/Tabla.java</file>
-            <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/McQuinnator/Fase2.java</file>
+            <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/gui/JFrameStarterPack.java</file>
             <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/McQuinnator/TablaPrimos.java</file>
-            <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/gui/JFrameConvertidora.java</file>
+            <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/McQuinnator/Fase2.java</file>
             <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/McQuinnator/Fase1.java</file>
             <file>file:/home/esteban/Descargas/Darkraft/Universidad/NetBeans/F0c%20Starter%20Pack/src/McQuinnator/Termino.java</file>
         </group>
